ide: Fix corruption of last data word during sector read
Don't increment the RAM address until after the read cycle
has ended.  Also don't squelch hidata when BSY is on because
the host will read that in the cycle after we see the last
word read internally.

Technically we oughtn't to set BSY and start the next operation
until after the host has read hidata, but hidata is latched and
won't be invalidated by other operations.