EtherZ: Add assembler implementations of mbuf copy-in/out
The C compiler makes a fairly poor job of this, and we can't use the
import/export routines from MBufManager because the address window for
the chip isn't big enough (we only have 0x200 bytes of data space
which clearly isn't enough for a 1500 byte packet).  Add assembler
versions instead.